perf(components): [el-scrollbar] prevent re-render when scrolling
