+++ packages/components/scrollbar/src/bar.ts
@@ -2,13 +2,25 @@ import { buildProps } from '@element-plus/utils/props'
 import type { ExtractPropTypes } from 'vue'
 
 export const barProps = buildProps({
-  vertical: Boolean,
-  size: String,
-  move: Number,
-  ratio: {
-    type: Number,
-    required: true,
+  always: {
+    type: Boolean,
+    default: true,
   },
-  always: Boolean,
+  width: {
+    type: String,
+    default: '',
+  },
+  height: {
+    type: String,
+    default: '',
+  },
+  ratioX: {
+    type: Number,
+    default: 1,
+  },
+  ratioY: {
+    type: Number,
+    default: 1,
+  },
 } as const)
 export type BarProps = ExtractPropTypes<typeof barProps>
