+++ packages/components/scrollbar/src/thumb.ts
@@ -1,0 +1,14 @@
+import { buildProps } from '@element-plus/utils/props'
+import type { ExtractPropTypes } from 'vue'
+
+export const thumbProps = buildProps({
+  vertical: Boolean,
+  size: String,
+  move: Number,
+  ratio: {
+    type: Number,
+    required: true,
+  },
+  always: Boolean,
+} as const)
+export type ThumbProps = ExtractPropTypes<typeof thumbProps>
